refactor: put user get and post logic in user/index.ts removed console.logs
